Use PSR-11 ContainerInterface
During registration of the flow, `ConvertMediaOperation` is fetched
from the container and handed to `registerOperation`. For some reason,
NC 27 could not find `ConvertMediaOperation` despite all paths and
namespaces looking good.

After comparing with the current version of the Automated PDF Converter
extension, I saw that there `ContainerInterface` was used instead of
`IServerContainer` to get the operation.

When doing this, `ConvertMediaOperation` is found, the flow is
registered correctly and the extension should work fine with NC 27.